+++ CHANGELOG.md
@@ -4,10 +4,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Added a `get_remaining_fragments` method to the `MolecularSystem` class that will return all remaining fragments. 
+
 ### Changed
 - Region options, '*x*\_program', '*x*\_method', '*x*\_xcfun', and '*x*\_basis', where *x* can be multipoles,
   polarizabilites, etc., have been unified to single options 'program', 'method', 'xcfun', and 'basis', that
   apply to all properties associated with a given region. All unused '*x*\_model' properties have been removed.
+
 ### Fixed
 - When running on a single local node, `ssh` is no longer required. This fixes problems with job processing
 - Process jobs without ssh when using only a single local machine. This fixes problems with job processing